% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install}
\alias{install}
\title{Install packages and their dependencies from CRAN-style repositories.}
\usage{
install(packages, lib_paths = .libPaths(), repos = repositories())
}
\arguments{
\item{packages}{character() vector of packages to install. Packages
and all dependencies must be available in the repositories
specified by \verb{repos = }.}

\item{lib_paths}{character() vector of (local) directory paths in
which to install packages (the first position) and to find
currently installed packages. \code{lib_paths} must have length at
least 1.}

\item{repos}{character() vector of paths (e.g., URLs) to
'CRAN-style' repositories. The default includes CRAN and
Bioconductor repositories relevant to the current R
installation.}
}
\value{
a tibble of package, version, dependency count (including
self), and logical value indicating whether installation
occurred.
}
\description{
Install packages and their dependencies from CRAN-style repositories.
}
